+++ MyRestaurant.SeedData.Tests/Fixture/DatabaseFixture.cs
@@ -1,13 +1,14 @@
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using MyRestaurant.Core;
 using System;
 
-namespace MyRestaurant.SeedData.Tests
+namespace MyRestaurant.SeedData.Tests.Fixture
 {
     public class DatabaseFixture : IDisposable
     {
         private readonly MyRestaurantContext _myRestaurantContext;
         public readonly string ConnectionString;
+        public readonly string WrongConnectionString;
         private readonly string _sereverName = "localhost";
         private readonly string _userName = "sa";
         private readonly string _password = "MyStr@ngPassw0rd";
@@ -16,13 +17,12 @@ namespace MyRestaurant.SeedData.Tests
         public DatabaseFixture()
         {
             ConnectionString = $"Server={_sereverName},1433;Database={Guid.NewGuid().ToString()};User={_userName};Password={_password}";
+            WrongConnectionString = $"Server=Test,1433;Database=Test;User=Test;Password=Test";
 
             var builder = new DbContextOptionsBuilder<MyRestaurantContext>();
 
             builder.UseSqlServer(ConnectionString);
             _myRestaurantContext = new MyRestaurantContext(builder.Options);
-
-            _myRestaurantContext.Database.Migrate();
         }
 
         public void Dispose()
